完善了TODO,操作逻辑更加完成

+++ library-server/src/main/java/com/firewood/controller/admin/TypeController.java
@@ -74,7 +74,7 @@ public class TypeController {
      */
     @DeleteMapping
     @Operation(summary = "类型批量删除")
-    public Result delete(@RequestParam List<Long> ids){
+    public Result delete(@RequestParam List<Integer> ids){
         log.info("类型批量删除:{}",ids);
         typeService.deleteBatch(ids);
         return Result.success();
